helper functions
One for getting the channel id from a user handle.
Another for getting the members playlist id from the channel id.

This is in preparation for making the program function as a quick way to get ids to pass to yt-dlp
to get json data of videos, which in turn will be fed to the program to determine which videos are missing
from an archive